CORDOPLUG-250 Extract applying the FCM Gradle plugin
If we leave the FCM gradle plugin initialization in the current plugin Johan always needs to add a 'google-services.json' file to his project, even if he doesn't want to use push mobile authentication. This is not very nice and therefore I created a separate plugin that just adds the FCM gradle plugin to an app project.